Reworking the javascript file to include image uploader logic
